Refactor functions into separate files
