+++ src/modules/__mocks__/tvmazeApiMock.js
@@ -1,37 +1,37 @@
 /* eslint-disable */
 const getShows = async () => {
-    return Promise.resolve([
-      {
-        id: 1,
-        image: {medium: '', original: ''},
-        name: "Under the Dome",
-        summary: "Under the Dome is the story of a",
-      },
-      {
-        id: 2,
-        image: {medium: '', original: ''},
-        name: "Under the Dome",
-        summary: "Under the Dome is the story of a",
-      },
-      {
-        id: 3,
-        image: {medium: '', original: ''},
-        name: "Under the Dome",
-        summary: "Under the Dome is the story of a",
-      },
-      {
-        id: 4,
-        image: {medium: '', original: ''},
-        name: "Under the Dome",
-        summary: "Under the Dome is the story of a",
-      },
-      {
-        id: 5,
-        image: {medium: '', original: ''},
-        name: "Under the Dome",
-        summary: "Under the Dome is the story of a",
-      },
-    ]);
+  return Promise.resolve([
+    {
+      id: 1,
+      image: { medium: '', original: '' },
+      name: 'Under the Dome',
+      summary: 'Under the Dome is the story of a',
+    },
+    {
+      id: 2,
+      image: { medium: '', original: '' },
+      name: 'Under the Dome',
+      summary: 'Under the Dome is the story of a',
+    },
+    {
+      id: 3,
+      image: { medium: '', original: '' },
+      name: 'Under the Dome',
+      summary: 'Under the Dome is the story of a',
+    },
+    {
+      id: 4,
+      image: { medium: '', original: '' },
+      name: 'Under the Dome',
+      summary: 'Under the Dome is the story of a',
+    },
+    {
+      id: 5,
+      image: { medium: '', original: '' },
+      name: 'Under the Dome',
+      summary: 'Under the Dome is the story of a',
+    },
+  ]);
 };
 
 export { getShows };
