+++ src/modules/getShowData.js
@@ -1,7 +1,7 @@
 const BASE_URL = 'https://api.tvmaze.com/shows';
 
 // fetch a single show by id
-const getShow = async (id) => {
+const getShowData = async (id) => {
   try {
     const result = await fetch(`${BASE_URL}/${id}`);
     const data = await result.json();
@@ -11,4 +11,4 @@ const getShow = async (id) => {
   }
 };
 
-export default getShow;
+export default getShowData;
